feat: user authorization for commands

+++ src/NadinSoftTask.Api/Controllers/ProductController.cs
@@ -1,28 +1,29 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Identity;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.SignalR;
+using NadinSoftTask.Application.Services;
+using NadinSoftTask.Core.Dtos.Account;
 using NadinSoftTask.Core.Dtos.Product;
 using NadinSoftTask.Core.Interfaces;
-using NadinSoftTask.Core.Models;
 using System.Security.Claims;
 
 namespace NadinSoftTask.Api.Controllers;
 
+[Authorize]
 [ApiController]
 [Route("[controller]")]
 public class ProductController : ControllerBase
 {
-    private readonly UserManager<ApplicationUser> _userManager;
     private readonly IProductService _productService;
+    private readonly IMapper _mapper;
 
-    public ProductController(IProductService productService, UserManager<ApplicationUser> userManager)
+    public ProductController(IProductService productService, IMapper mapper)
     {
         _productService = productService;
-        _userManager = userManager;
+        _mapper = mapper;
     }
 
+    [AllowAnonymous]
     [HttpGet]
     [Route("Get/{id}")]
     public async Task<IActionResult> GetProductById(int id)
@@ -31,6 +32,7 @@ public class ProductController : ControllerBase
         return Ok(product);
     }
 
+    [AllowAnonymous]
     [HttpGet]
     [Route("Get")]
     public async Task<IActionResult> GetProducts([FromQuery] ProductFilterDto productFilter)
@@ -39,29 +41,45 @@ public class ProductController : ControllerBase
         return Ok(products);
     }
 
-    [Authorize]
     [HttpPost]
     [Route("Create")]
-    public async Task<IActionResult> CreateProduct(ProductCreateDto productCreate)
+    public async Task<IActionResult> CreateProduct(ProductDetailsDto productDetails)
     {
+        var userIdDto = new UserIdDto
+        (
+            UserId: User.FindFirstValue(ClaimTypes.NameIdentifier) ?? throw new NullReferenceException()
+        );
+
+        ProductCreateDto productCreate = _mapper.Map<ProductCreateDto>(productDetails);
+        _mapper.Map<UserIdDto, ProductCreateDto>(userIdDto, productCreate);
+
         int id = await _productService.CreateProductAsync(productCreate);
         return Ok(id);
     }
 
-    [Authorize]
     [HttpPut]
     [Route("Update")]
     public async Task<IActionResult> UpdateProduct(ProductUpdateDto productUpdate)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        
+        if (userId != await _productService.GetCreatorId(productUpdate.Id))
+            return Unauthorized();
+
         ProductGetDto updatedProduct = await _productService.UpdateProductAsync(productUpdate);
+        
         return Ok(updatedProduct);
     }
 
-    [Authorize]
     [HttpDelete]
     [Route("Delete")]
     public async Task<IActionResult> DeleteProduct(int id)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+        if (userId != await _productService.GetCreatorId(id))
+            return Unauthorized();
+
         await _productService.DeleteProductAsync(id);
         return Ok();
     }
